feat: 기존의 Default와 Multiple 이외의 Disabled와 CollapsibleGroup을 추가.
